+++ op_tests/test_mla.py
@@ -486,7 +486,7 @@ def test_mla(
         return err, us_asm_decode
 
     def test_absorb_decode_gluon():
-        from aiter.ops.triton._gluon_kernels.gfx950.attention.mla import mla_gluon
+        from aiter.ops.triton.gluon.mla_gluon import mla_gluon
 
         out_gluon = torch.empty((total_q, nhead, v_head_dim), dtype=out_dtype).fill_(-1)
 
@@ -539,7 +539,7 @@ def test_mla(
         # Shared bh16bn{64,128} runner. The wrapper dispatches on
         # (nhead, kv dtype): name='bh16bn128' -> cast kv to fp8;
         # name='bh16bn64' -> keep bf16. -lse also validates the returned lse.
-        from aiter.ops.triton._gluon_kernels.gfx950.attention.mla import mla_gluon
+        from aiter.ops.triton.gluon.mla_gluon import mla_gluon
 
         out_gluon = torch.empty((total_q, nhead, v_head_dim), dtype=out_dtype).fill_(-1)
         # MTP (decode_qlen>1): q is [total_q=batch*qlen, nhead, qk]; reshape to
